+++ src/OpenTelemetry.OpAmp.Client/CHANGELOG.md
@@ -5,6 +5,9 @@
 * Assemblies are now digitally signed using cosign.
   ([#4637](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4637))
 
+* Add support for reporting full state.
+  ([#4633](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4633))
+
 ## 0.5.0-alpha.1
 
 Released 2026-Jul-01